Refactor AWS dependencies. Move cloudwatch lib module to more general AWS package. Consolidate AWS initializers into AwsInitializers. Cleanup injectors accross the board. Adjusted build config.
