+++ server/backend-shared/src/main/scala/cool/graph/deprecated/actions/schemas/ActionUserContext.scala
@@ -2,8 +2,8 @@ package cool.graph.deprecated.actions.schemas
 
 import akka.actor.ActorRef
 import cool.graph.RequestContextTrait
+import cool.graph.aws.cloudwatch.Cloudwatch
 import cool.graph.client.database.ProjectDataresolver
-import cool.graph.cloudwatch.Cloudwatch
 import cool.graph.shared.models.Project
 import scaldi.{Injectable, Injector}
