+++ server/backend-shared/src/main/scala/cool/graph/subscriptions/SubscriptionUserContext.scala
@@ -1,10 +1,10 @@
 package cool.graph.subscriptions
 
 import cool.graph.RequestContextTrait
+import cool.graph.aws.cloudwatch.Cloudwatch
 import cool.graph.client.UserContextTrait
 import cool.graph.deprecated.actions.schemas.MutationMetaData
 import cool.graph.client.database.ProjectDataresolver
-import cool.graph.cloudwatch.Cloudwatch
 import cool.graph.shared.models.{AuthenticatedRequest, Project}
 import sangria.ast.Document
 import scaldi.{Injectable, Injector}
